requirements.txt
streamlit
pandas

langchain>=0.2.11
langchain-core>=0.2.25
langchain-openai>=0.1.17
langchain-community>=0.2.11

faiss-cpu
tavily-python